Scripts/requirements.txt
# Core libraries
langchain==0.0.143
transformers==4.30.2
#torch  # Updated to the latest stable version
sentence-transformers==2.2.2
faiss-cpu==1.7.3  # For vector storage and similarity search

# Document loaders (PDF, text, etc.)
pypdf2==3.0.1  # For loading PDFs
python-docx==0.8.11  # For loading Word documents
beautifulsoup4==4.12.2  # For HTML parsing
tika==1.24  # Apache Tika for various document formats

# For Elasticsearch integration (optional, if needed)
elasticsearch==8.9.0
elasticsearch-dsl==8.9.0

# Weights & Biases (for tracking models, optional)
wandb==0.15.8

# Additional utilities
numpy==1.24.3  # For array manipulations
scikit-learn==1.3.0  # For additional machine learning utilities
